tests/wpt/web-platform-tests/html/semantics/embedded-content/the-canvas-element/initial.colour.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>Canvas test: initial.colour</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>
<link rel="stylesheet" href="/common/canvas-tests.css">
<body class="show_output">

<h1>initial.colour</h1>
<p class="desc">Initial state is transparent black</p>

<p class="notes">Output should be transparent black (not transparent anything-else), but manual
verification can only confirm that it's transparent - it's not possible to make
the actual blackness visible.

<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="initial.colour.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
var t = async_test("Initial state is transparent black");
_addTest(function(canvas, ctx) {

_assertPixel(canvas, 20,20, 0,0,0,0, "20,20", "0,0,0,0");


});
</script>
